perf: system init
